# src/main/resources/apidocs.json
{  
    "swagger": "2.0",
    "info": {  
        "description": "TestbedRegistry API documentation",
        "version": "0.0.1",
        "title": "TestbedRegistry API",
        "contact": {  

        },
        "license": {  

        }
    },
    "host": "platform-dev.fiesta-iot.eu",
    "basePath": "/ui.testbed-registry",
...
    "paths": {  
...
        "/api/register-testbeds": {  

            "post": {  
                "tags": [  
                    "register-testbeds-resource"
                ],
                "summary": "createRegisterTestbeds",
                "operationId": "createRegisterTestbedsUsingPOST",
                "consumes": [  
                    "application/json"
                ],
                "produces": [  
                    "*/*"
                ],
                "parameters": [  
                    {  
                        "in": "body",
                        "name": "registerTestbeds",
                        "description": "registerTestbeds",
                        "required": true,
                        "schema": {  
                            "$ref": "#/definitions/RegisterTestbedsDTO"
                           
                        }
                    }
                ],
                "responses": {  
                    
                    "201": {  
                        "description": "Created, register testbed success",                       
                        "schema": {  
                            "type": "object"
                        }
                    },
                    "401": {  
                        "description": "Unauthorized"
                    },
                    "403": {  
                        "description": "Forbidden"
                    },
                    "404": {  
                        "description": "Not Found"
                    },
                   "500": {
                        "description": "Internal Server Error, Response body error {\"errorMessage\":\"error message conent here\"} "     
                    }    
                }
            }
        },
        "/api/register-testbeds/validateAnnotatedObseration": {  
            "post": {  
                "tags": [  
                    "register-testbeds-resource"
                ],
                "summary": "validateAnnotatedObseration",
                "operationId": "validateAnnotatedObserationUsingPOST",
                "consumes": [  
                    "application/json"
                ],
                "produces": [  
                    "*/*"
                ],
                "parameters": [  
                    {  
                        "in": "body",
                        "name": "ontologyValidatorBody",
                        "description": "ontologyValidatorBody",
                        "required": true,
                        "schema": {  
                            "$ref": "#/definitions/OntologyValidatorBody"
                        }
                    }
                ],
                "responses": {  
                    "200": {  
                        "description": "OK",
                        "schema": {  
                            "$ref": "#/definitions/OntologyValidatorResponse"
                        }
                    },
                    
                    "401": {  
                        "description": "Unauthorized"
                    },
                    "403": {  
                        "description": "Forbidden"
                    },
                    "404": {  
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error, Response body error {\"errorMessage\":\"error message conent here\"} "     
                    }
                }
            }
        },
        "/api/register-testbeds/validateAnnotatedResource": {  
            "post": {  
                "tags": [  
                    "register-testbeds-resource"
                ],
                "summary": "validateAnnotatedResource",
                "operationId": "validateAnnotatedResourceUsingPOST",
                "consumes": [  
                    "application/json"
                ],
                "produces": [  
                    "*/*"
                ],
                "parameters": [  
                    {  
                        "in": "body",
                        "name": "body",
                        "description": "body",
                        "required": true,
                        "schema": {  
                            "$ref": "#/definitions/OntologyValidatorBody"
                        }
                    }
                ],
                "responses": {  
                    "200": {  
                        "description": "OK",
                        "schema": {  
                            "$ref": "#/definitions/OntologyValidatorResponse"
                        }
                    },
                    
                    "401": {  
                        "description": "Unauthorized"
                    },
                    "403": {  
                        "description": "Forbidden"
                    },
                    "404": {  
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error, Response body error {\"errorMessage\":\"error message conent here\"} "     
                    }
                }
            }
        },
...
        "/api/getTestbedByID/{id}": {  
            "get": {  
                "tags": [  
                    "testbed-resource"
                ],
                "summary": "getRegisterTestbeds",
                "operationId": "getRegisterTestbedsUsingGET_1",
                "consumes": [  
                    "application/json"
                ],
                "produces": [  
                    "*/*"
                ],
                "parameters": [  
                    {  
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {  
                    "200": {  
                        "description": "OK",
                        "schema": {  
                            "$ref": "#/definitions/RegisterTestbeds"
                        }
                    },
                    "401": {  
                        "description": "Unauthorized"
                    },
                    "403": {  
                        "description": "Forbidden"
                    },
                    "404": {  
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error, Response body error {\"errorMessage\":\"error message conent here\"} "     
                    }
                }
            }
        },
...
    },
    "definitions": {  
        "RegisterTestbedsDTO": {  
            "type": "object",
            "required": [  
                "annotatedObservation",
                "annotatedResourceDescription",
                "iri",
                "latitude",
                "longitude",
                "name",
                "resourceDescriptionContentType",
                "resourceObservationContentType"
            ],
            "properties": {  
                "annotatedObservation": {  
                    "type": "string"
                },
                "annotatedResourceDescription": {  
                    "type": "string"
                },
                "getApiKey": {  
                    "type": "string"
                },
                "getLastObservationsURL": {  
                    "type": "string"
                },
                "getObservationsURL": {  
                    "type": "string"
                },
                "id": {  
                    "type": "integer",
                    "format": "int64"
                },
                "iri": {  
                    "type": "string"
                },
                "latitude": {  
                    "type": "string"
                },
                "longitude": {  
                    "type": "string"
                },
                "name": {  
                    "type": "string"
                },
                "pushApiKey": {  
                    "type": "string"
                },
                "pushLastObservationsURL": {  
                    "type": "string"
                },
                "pushObservationsURL": {  
                    "type": "string"
                },
                "registerID": {  
                    "type": "string"
                },
                "resourceDescriptionContentType": {  
                    "type": "string"
                },
                "resourceID": {  
                    "type": "string"
                },
                "resourceObservationContentType": {  
                    "type": "string"
                },
                "resourceType": {  
                    "type": "string"
                },
                "userID": {  
                    "type": "string"
                }
            }
        },
        "OntologyValidatorBody": {  
            "type": "object",
            "properties": {  
                "annotatedObservation": {  
                    "type": "string"
                },
                "annotatedResourceDescription": {  
                    "type": "string"
                },
                "contentType": {  
                    "type": "string"
                }
            }
        },
...
        "OntologyValidatorResponse": {  
            "type": "object",
            "properties": {  
                "message": {  
                    "type": "string"
                },
                "result": {  
                    "type": "boolean"
                }
            }
        },
...
        "RegisterTestbeds": {  
            "type": "object",
            "required": [  
                "annotatedObservation",
                "annotatedResourceDescription",
                "iri",
                "latitude",
                "longitude",
                "name"
            ],
            "properties": {  
                "annotatedObservation": {  
                    "type": "string"
                },
                "annotatedResourceDescription": {  
                    "type": "string"
                },
                "getApiKey": {  
                    "type": "string"
                },
                "getLastObservationsURL": {  
                    "type": "string"
                },
                "getObservationsURL": {  
                    "type": "string"
                },
                "id": {  
                    "type": "integer",
                    "format": "int64"
                },
                "iri": {  
                    "type": "string"
                },
                "latitude": {  
                    "type": "string"
                },
                "longitude": {  
                    "type": "string"
                },
                "name": {  
                    "type": "string"
                },
                "pushApiKey": {  
                    "type": "string"
                },
                "pushLastObservationsURL": {  
                    "type": "string"
                },
                "pushObservationsURL": {  
                    "type": "string"
                },
                "registerID": {  
                    "type": "string"
                },
                "resourceID": {  
                    "type": "string"
                },
                "resourceType": {  
                    "type": "string"
                },
                "userID": {  
                    "type": "string"
                }
            }
        },
...
    }
}